IntermediateTextInputUIView
